+++ src/hr/avrbanac/openglplayground/loaders/ModelLoader.java
@@ -1,9 +1,11 @@
 
 package hr.avrbanac.openglplayground.loaders;
 
+import de.matthiasmann.twl.utils.PNGDecoder;
 import hr.avrbanac.openglplayground.models.RawModel;
 import static hr.avrbanac.openglplayground.utils.BufferUtils.*;
 import static hr.avrbanac.openglplayground.Globals.*;
+import hr.avrbanac.openglplayground.textures.TextureData;
 import java.awt.image.BufferedImage;
 import java.nio.FloatBuffer;
 import java.nio.IntBuffer;
@@ -15,14 +17,17 @@ import org.lwjgl.opengl.GL20;
 import org.lwjgl.opengl.GL30;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.nio.ByteBuffer;
 import javax.imageio.ImageIO;
+import org.lwjgl.opengl.GL12;
+import org.lwjgl.opengl.GL13;
 import org.lwjgl.opengl.GL14;
 
 /**
  * Loads a model into memory by storing data into VAO.
  * 
  * @author avrbanac
- * @version 1.0.7
+ * @version 1.0.14
  */
 public class ModelLoader {
     
@@ -45,12 +50,65 @@ public class ModelLoader {
         return new RawModel(vaoID, indices.length);
     }
     
-    public RawModel loadToVAO(float[] positions) {
+    public RawModel loadToVAO(float[] positions, int dimensions) {
         int vaoID = createVAO();
-        storeDataInAttributeList(0, 2, positions);
+        storeDataInAttributeList(0, dimensions, positions);
         unbindVAO();
         
-        return new RawModel(vaoID, positions.length/2);
+        return new RawModel(vaoID, positions.length / dimensions);
+    }
+    
+    public int loadCubeMap(String[] textureFiles) {
+        int texID = GL11.glGenTextures();
+        GL13.glActiveTexture(GL13.GL_TEXTURE0);
+        GL11.glBindTexture(GL13.GL_TEXTURE_CUBE_MAP, texID);
+        for(int i=0; i<textureFiles.length; i++) {
+            TextureData data = decodeTextureFile(TEXTURE_CUBE_FILE + textureFiles[i] + TEXTURE_CUBE_EXT);
+            
+            // GL_TEXTURE_CUBE_MAP_POSITIVE_X = 34069 -> right face
+            // GL_TEXTURE_CUBE_MAP_NEGATIVE_X = 34070 -> left face
+            // GL_TEXTURE_CUBE_MAP_POSITIVE_X = 34071 -> top face
+            // GL_TEXTURE_CUBE_MAP_NEGATIVE_X = 34072 -> bottom face
+            // GL_TEXTURE_CUBE_MAP_POSITIVE_X = 34073 -> back face
+            // GL_TEXTURE_CUBE_MAP_NEGATIVE_X = 34074 -> front face
+            GL11.glTexImage2D(
+                    GL13.GL_TEXTURE_CUBE_MAP_POSITIVE_X + i,
+                    0, GL11.GL_RGBA, data.getWidth(), data.getHeight(),
+                    0, GL11.GL_RGBA, GL11.GL_UNSIGNED_BYTE, data.getBuffer());
+        }
+        // make texture appear smooth
+        GL11.glTexParameteri(GL13.GL_TEXTURE_CUBE_MAP, GL11.GL_TEXTURE_MAG_FILTER, GL11.GL_LINEAR);
+        GL11.glTexParameteri(GL13.GL_TEXTURE_CUBE_MAP, GL11.GL_TEXTURE_MIN_FILTER, GL11.GL_LINEAR);
+        
+        // fix the visible seams at the edges of the skybox (hardwere limitation on some computers)
+        if (FIX_SKYBOX_SEAMS) {
+            GL11.glTexParameteri(GL13.GL_TEXTURE_CUBE_MAP, GL11.GL_TEXTURE_WRAP_S, GL12.GL_CLAMP_TO_EDGE);
+            GL11.glTexParameteri(GL13.GL_TEXTURE_CUBE_MAP, GL11.GL_TEXTURE_WRAP_T, GL12.GL_CLAMP_TO_EDGE);
+        }
+        
+        texs.add(texID);
+        return texID;
+    }
+    
+    public TextureData decodeTextureFile(String fileName) {
+        int width = 0;
+        int height = 0;
+        ByteBuffer buffer = null;
+        
+        try (FileInputStream in = new FileInputStream(fileName)) {
+            PNGDecoder decoder = new PNGDecoder(in);
+            width   = decoder.getWidth();
+            height  = decoder.getHeight();
+            buffer  = ByteBuffer.allocateDirect(4 * width * height);
+            decoder.decode(buffer, width * 4, PNGDecoder.Format.RGBA);
+            buffer.flip();
+        } catch (IOException e) {
+            e.printStackTrace();
+            System.err.println("Tried to load texture " + fileName + ", didn't work");
+            System.exit(-1);
+        }
+        
+        return new TextureData(width, height, buffer);
     }
     
     private int createVAO() {